Fixed mulit-detector summing and updated data urls
Summing of detectors would fail if detector 0 was not specified due to detector 0 being hardcoded into the edges_over_timespan function call arguments. This has now been changed to the first detector in the list of detectors.

I have also changed the data urls to allow me to run the unit tests. It is possible the original ones are suitable and should not be changed but for me it did not appear to work with the ftp urls.